Removed old c examples, renamed plain_interface example to asteroid
