--- examples/simplest/holman_ic
@@ -1,8 +1,0 @@
-tstart 7304.5
-tepoch 2458849.5
-tend   7404.5
-tstep 20.0
-geocentric 0
-state
- 3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
- 2.8056633153049852e-03  7.5504086883996860e-03  2.9800282074358684e-03
