--- examples/simplest/test_ic_v2
@@ -1,25 +1,0 @@
-tstart 2458849.5
-tepoch 2458849.5
-tend   2458849.5
-tstep 20.0
-geocentric 0
-state
- 3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
- 2.8056633153049852e-03  7.5504086883996860e-03  2.9800282074358684e-03
-A1A2A3
-1e-3 0.0 0.9
-var 0
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-A1A2A3
-1.0 0.0 0.0
-var 0
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-A1A2A3
-0.0 1.0 0.0
-var 0
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-A1A2A3
-0.0 0.0 1.0
