--- examples/simplest/test_var_ic_2nd
@@ -1,62 +1,0 @@
-tstart 2458849.5
-tepoch 2458849.5
-tend   2458849.5
-tstep 20.0
-geocentric 0
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753402614091e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765183678903173e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385907775843308e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056533153049852e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5503986883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800182074358685e-03
-state
-3.3388753602614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765181678903163e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385905775843298e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056733153049851e-03 7.5504086883996860e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5504186883996859e-03 2.9800282074358684e-03
-state
-3.3388753502614090e+00 -9.1765182678903168e-01 -5.0385906775843303e-01
-2.8056633153049852e-03 7.5504086883996860e-03 2.9800382074358684e-03
-var 0
-1.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-var 0
-0.0000000000000000e+00 1.0000000000000000e+00 0.0000000000000000e+00
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-var 0
-0.0000000000000000e+00 0.0000000000000000e+00 1.0000000000000000e+00
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-var 0
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-1.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-var 0
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-0.0000000000000000e+00 1.0000000000000000e+00 0.0000000000000000e+00
-var 0
-0.0000000000000000e+00 0.0000000000000000e+00 0.0000000000000000e+00
-0.0000000000000000e+00 0.0000000000000000e+00 1.0000000000000000e+00
