--- examples/tests/acc_eih_sun_only.out
@@ -1,31 +1,0 @@
-  0    2.4588495000000000e+06    2.9591220828411962e-04    3.3426746816397150e+00   -9.2446835233552804e-01   -5.0683959610231433e-01   -2.2971753123485163e-05    6.3531934103468973e-06    3.4831370635080606e-06
-  1    2.4588495000000000e+06    4.9125001948893182e-11    3.4060095581191883e+00   -5.1431928362203960e-01   -2.9430624045878817e-01   -4.0493498406679731e-12    6.1146590273734408e-13    3.4989594350036774e-13
-  2    2.4588495000000000e+06    7.2434523326441187e-10    2.6194743816042498e+00   -9.8899448686953162e-01   -4.9011366693491193e-01   -8.2612418798156486e-11    3.1190695092156604e-11    1.5457099254671876e-11
-  3    2.4588495000000000e+06    8.8876924467071022e-10    3.5090204418661886e+00   -1.8136353387573383e+00   -8.9229569851278034e-01   -4.6963581887223305e-11    2.4273102182327000e-11    1.1942193782842626e-11
-  4    2.4588495000000000e+06    1.0931894624024351e-11    3.5064122119661620e+00   -1.8131238168018737e+00   -8.9182293340072682e-01   -5.7833480437613407e-13    2.9905001024161975e-13    1.4709401249696446e-13
-  5    2.4588495000000000e+06    9.5495488297258119e-11    4.6627822869364079e+00   -1.0630087749190875e-01   -1.6719343985598262e-01   -4.3804374668023679e-12    9.9864054949361225e-14    1.5706939828621318e-13
-  6    2.4588495000000000e+06    2.8253458252257917e-07    2.8165275615052030e+00    3.8512875158142141e+00    1.5529893487966482e+00   -6.2991095114555524e-09   -8.6133301707335444e-09   -3.4732306943824877e-09
-  7    2.4588495000000000e+06    8.4597059933762903e-08   -4.5456817122150595e-01    7.6013036421499471e+00    3.1782723722812611e+00    6.8446260022529563e-11   -1.1445605709759853e-09   -4.7856596873250915e-10
-  8    2.4588495000000000e+06    1.2920265649682399e-08   -1.2882822619647998e+01   -1.1431312561706914e+01   -4.8791258021774038e+00    2.9016587607485993e-11    2.5747283200922229e-11    1.0989484639096455e-11
-  9    2.4588495000000000e+06    1.5243573478851939e-08   -2.5900138514828157e+01    4.7013600221270799e+00    2.5239830341250022e+00    2.1350226069907247e-11   -3.8754657335508562e-12   -2.0805915128341060e-12
- 10    2.4588495000000000e+06    2.1750964648933581e-12   -9.6340467513988752e+00    2.7699344337300907e+01    1.2335328102384008e+01    6.5068261505687606e-16   -1.8708111216232986e-15   -8.3312690444213981e-16
- 11    2.4588495000000000e+06    3.2191392075878588e-15    5.1673878230710351e+00   -3.5460450404562489e+00   -1.1792883639362692e+00   -6.4142628993615471e-17    4.4016949997273036e-17    1.4638470847248433e-17
- 12    2.4588495000000000e+06    1.3964518123081070e-13    2.3350658125947739e+00    1.4655959231199995e+00    8.2528492664419584e-01   -1.3682132284300127e-14   -8.5875426668064925e-15   -4.8356913444068153e-15
- 13    2.4588495000000000e+06    2.0917175955133682e-15    6.5938166112906824e+00   -5.6878357004313829e-01   -4.6495178291490014e-01   -4.7227264514438608e-17    4.0738306351889865e-18    3.3301503715742043e-18
- 14    2.4588495000000000e+06    8.6836253492286545e-15    4.2714710338689681e+00   -3.1464809599723198e+00   -1.5161831220796609e+00   -2.2080695019457406e-16    1.6265236474926691e-16    7.8376692354544792e-17
- 15    2.4588495000000000e+06    4.5107799051436795e-15    1.1913867200927255e+00   -8.4171568608710767e-01   -9.2836618475361765e-01   -1.0395653083415707e-15    7.3445373528670711e-16    8.1006214256960051e-16
- 16    2.4588495000000000e+06    2.4067012218937576e-15    5.3047519743048914e+00    1.5364816196557363e+00    1.4059135646415717e+00   -6.8977605575393605e-17   -1.9978846070054536e-17   -1.8281071726760478e-17
- 17    2.4588495000000000e+06    5.9824315264869841e-15    5.2478514218313510e-01   -2.1060979035145286e+00   -6.0384149473055948e-01   -2.7454838601881548e-16    1.1018333670844294e-15    3.1590777722820966e-16
- 18    2.4588495000000000e+06    1.2542530761640810e-14    2.0760503444989604e+00   -3.8467737843372545e+00   -1.9178797638526448e+00   -2.3939953761096240e-16    4.4359033378189445e-16    2.2116011294059196e-16
- 19    2.4588495000000000e+06    6.3110343420878887e-15    6.6170824658588714e+00   -1.8791501390601606e+00    1.3604622599217064e-01   -1.2822913637463933e-16    3.6415112051757075e-17   -2.6363718687221692e-18
- 20    2.4588495000000000e+06    2.5416014973471498e-15    4.6830193680947989e+00    1.3973015680041172e+00    5.8970796273622972e-01   -9.9788316795196068e-17   -2.9774459716391346e-17   -1.2565817131376284e-17
- 21    2.4588495000000000e+06    4.2823439677995011e-15    6.1889599347987652e+00   -1.0246886598927099e+00   -6.4049027409458892e-01   -1.0569958945492138e-16    1.7500383232533349e-17    1.0938761881622906e-17
- 22    2.4588495000000000e+06    3.0471146330043200e-14    3.9597442874684052e+00    2.0750796988448763e+00   -1.1439697606311308e+00   -1.2279028867721233e-15   -6.4347497401729188e-16    3.5474103110760241e-16
- 23    2.4588495000000000e+06    3.5445002842488978e-15    8.1518865903801174e-01   -5.0639951837070918e-01   -2.7545932958027697e-01   -2.9031392005523275e-15    1.8034454682646320e-15    9.8099595596198797e-16
- 24    2.4588495000000000e+06    4.8345606546105521e-15    4.8496692902934297e+00   -3.8199095186034095e+00   -2.2691170343561664e+00   -8.2402611384921485e-17    6.4905563811753043e-17    3.8555447387556486e-17
- 25    2.4588495000000000e+06    2.6529436610356353e-15    2.4502564269567162e+00   -3.4393784363479374e+00   -1.7165877027301710e+00   -6.8624463737320359e-17    9.6326857135206291e-17    4.8076564257499901e-17
- 26    2.4588495000000000e+06    3.8548000225257904e-14    2.0990691533635251e+00   -3.0583387229025107e+00   -1.1943262710992117e+00   -1.3673068244312746e-15    1.9921627644074246e-15    7.7796887179957415e-16
-J24    2.4588495000000000e+06   -6.4016432868096248e-24    3.3086859091940721e-24    5.9272835926745608e-24
- J2    2.4588495000000000e+06   -1.3023962309511366e-17    4.6734608690892927e-18   -5.4774520501613907e-19
- GR    2.4588495000000000e+06    2.0454531796156866e-13   -4.9590257762643515e-14   -2.8159733950256379e-14
- GR    2.4588495000000000e+06    2.0476033572473280e-13   -4.9784786902228314e-14   -2.8249668305991862e-14
