--- examples/tests/eih_acc.out
@@ -1,22 +1,0 @@
- EIH_J           0
-  -7.3754593879144891E-06   2.6413247162140478E-04   6.0675859299828608E-09  -1.6780846781637474E-09  -9.2000960158050587E-10   3.6291911097832037E-03   7.0183716084583260E-11   1.8836085773837614E-10   7.4338682126789094E-11   7.5352722701818583E-13  -2.7774835756718876E-12  -1.2276896955501858E-12   2.5500907742631091E-09  -9.3995744122159685E-09  -4.1547538748787059E-09
- EIH_J           1
-   2.2342452332312395E-02   7.2824353546175333E-04   2.9489128442895321E-15  -4.4529609082375604E-16  -2.5480945893843367E-16  -2.1254211689439215E-01   4.9057212667247704E-15  -2.2400871492001114E-15  -1.5125839340852190E-15   9.1951767608431237E-15   5.9545483674852413E-14   3.0855667966467080E-14   1.8488875807374483E-04   1.1972897108869117E-03   6.2041940880893298E-04
- EIH_J           2
-  -9.2275996534141209E-03  -8.8465536293869462E-05  -7.3083519335126566E-15   2.7593015687061957E-15   1.3674205751121176E-15   7.2375133657454682E-02   9.9548481565086286E-15  -2.4484491807260546E-14  -1.2193417202934158E-14  -4.9824607942705921E-13  -4.4463767977889388E-14   1.1519568003110439E-14  -5.5864388228661188E-04  -4.9853702798469545E-05   1.2915979587681534E-05
- EIH_J           3
-  -1.3458009377975186E-02   3.3201458535950614E-04   1.5592594167283663E-14  -8.0590239564542247E-15  -3.9649825170933674E-15   1.3746532041413151E-01   3.6893712355700129E-14   1.8927747083956847E-14   7.6649357366718923E-15   4.0886370805874944E-14  -2.1281755830609113E-13  -9.2364081213385715E-14   5.3226274407648183E-05  -2.7704796326745716E-04  -1.2024045752105648E-04
- EIH_J           4
-  -1.3600571884643616E-02   1.3291989968061644E-04   7.6872204179484685E-17  -3.9749697360803420E-17  -1.9551721384715860E-17   1.3908325584879358E-01   4.5672292598200297E-16   2.2444611507999463E-16   9.1066643447599201E-17  -6.3047884554666271E-16  -2.4053710442036911E-15  -9.3478747714849365E-16  -6.6685804896800038E-05  -2.5441631434770757E-04  -9.8872556567772594E-05
- EIH_J           5
-   8.6814724484812495E-03   1.8439215874301233E-04   8.0771832074246104E-16  -1.8414148672943520E-17  -2.8962365422460849E-17  -7.4423661696316437E-02   3.8502220226199858E-16  -1.1848818880662650E-15  -5.2537421633390662E-16   6.9629048790028454E-15   4.3143581056785960E-15   1.7910047891173053E-15   9.7224652548846125E-05   6.0242380886290427E-05   2.5008214439400246E-05
- EIH_J           6
-   5.0887124046125045E-03   3.8755779598690482E-04   2.4412689989398626E-12   3.3381632570770029E-12   1.3460776328689440E-12   8.9835450427531424E-02  -9.2628320339486858E-13   1.2978614395580394E-12   5.4107008906956571E-13  -2.1439960717176132E-13   1.9491570344084517E-12   8.4068890650517845E-13  -1.0878937478233356E-06   9.8902968117850060E-06   4.2657736985437917E-06
- EIH_J           7
-   1.8267275721562651E-03   3.7694095940709937E-04  -2.5800198920720082E-14   4.3143175972322533E-13   1.8039111539361992E-13   2.1713576954972816E-01  -6.7004903028908154E-14   1.8104291246608188E-13   7.7082870081550849E-14  -4.0137572825887191E-14   8.9868548617298599E-14   3.8844352710840309E-14  -1.1185686650125866E-06   2.5044898177957639E-06   1.0825287304622232E-06
- EIH_J           8
-  -4.6069971388918422E-04   3.3002838374258112E-04  -9.5762975098236118E-15  -8.4973342605628747E-15  -3.6268418536049248E-15  -5.2324247056364559E-01  -6.0036422729217841E-15  -5.6348696678953357E-15  -2.0445804766718658E-15  -1.5592168472863633E-15  -1.0102356860488807E-15  -4.2040533898326214E-16  -6.1722817466729893E-07  -3.9990969156658783E-07  -1.6642073900922203E-07
- EIH_J           9
-  -1.3833207293087929E-05   3.6607732021977330E-04  -7.8158335457579865E-15   1.4187201103418555E-15   7.6165736549031365E-16  -1.1749593301702177E-01  -2.0914038379558005E-16  -4.5403758831260574E-16  -1.7670619613063522E-16  -6.5195481932820447E-16   1.2526726133306647E-16   6.7504052312475681E-17  -3.2314055476263031E-07   6.2088554483688468E-08   3.3458295370001474E-08
- EIH_J          10
-  -4.0183313337176018E-04   3.0895044349114230E-04  -2.0102868249379807E-19   5.7798792571367943E-19   2.5739492661180161E-19   9.1383803661908847E-01  -1.0269177216365395E-20   4.1340067653072546E-19   2.2277874851025068E-19  -2.3525758205104483E-20   5.1892338309511111E-20   2.3282022067309778E-20  -9.8318765845070388E-08   2.1686827752480248E-07   9.7300144636312919E-08
