--- examples/tests/jpl.out
@@ -1,22 +1,0 @@
- EIH_J           0
-  -7.3754593879144899E-06   2.6413247162140478E-04   6.0675859299828608E-09  -1.6780846781637474E-09  -9.2000960158050587E-10   3.6291911097832037E-03   7.0183716084583260E-11   1.8836085773837614E-10   7.4338682126789094E-11   7.5352722701818573E-13  -2.7774835756718868E-12  -1.2276896955501851E-12   2.5500907742631091E-09  -9.3995744122159669E-09  -4.1547538748787042E-09
- EIH_J           1
-   2.2342452332312392E-02   7.2824353546175344E-04   2.9489128442895332E-15  -4.4529609082375614E-16  -2.5480945893843372E-16  -2.1254211689439206E-01   4.9057212667247664E-15  -2.2400871492001102E-15  -1.5125839340852187E-15   9.1951767608431222E-15   5.9545483674852425E-14   3.0855667966467099E-14   1.8488875807374481E-04   1.1972897108869117E-03   6.2041940880893330E-04
- EIH_J           2
-  -9.2275996534141227E-03  -8.8465536293869801E-05  -7.3083519335126850E-15   2.7593015687062064E-15   1.3674205751121227E-15   7.2375133657454696E-02   9.9548481565086270E-15  -2.4484491807260558E-14  -1.2193417202934168E-14  -4.9824607942705931E-13  -4.4463767977889413E-14   1.1519568003110447E-14  -5.5864388228661199E-04  -4.9853702798469579E-05   1.2915979587681544E-05
- EIH_J           3
-  -1.3458009377975189E-02   3.3201458535950641E-04   1.5592594167283672E-14  -8.0590239564542278E-15  -3.9649825170933697E-15   1.3746532041413162E-01   3.6893712355700154E-14   1.8927747083956863E-14   7.6649357366718955E-15   4.0886370805874988E-14  -2.1281755830609128E-13  -9.2364081213385790E-14   5.3226274407648238E-05  -2.7704796326745733E-04  -1.2024045752105658E-04
- EIH_J           4
-   8.6814724484812478E-03   1.8439215874301238E-04   8.0771832074246133E-16  -1.8414148672943585E-17  -2.8962365422460849E-17  -7.4423661696316423E-02   3.8502220226199839E-16  -1.1848818880662648E-15  -5.2537421633390682E-16   6.9629048790028486E-15   4.3143581056785960E-15   1.7910047891173061E-15   9.7224652548846152E-05   6.0242380886290427E-05   2.5008214439400263E-05
- EIH_J           5
-   5.0887124046125019E-03   3.8755779598690482E-04   2.4412689989398614E-12   3.3381632570770021E-12   1.3460776328689432E-12   8.9835450427531494E-02  -9.2628320339486818E-13   1.2978614395580396E-12   5.4107008906956571E-13  -2.1439960717176120E-13   1.9491570344084509E-12   8.4068890650517815E-13  -1.0878937478233352E-06   9.8902968117850043E-06   4.2657736985437909E-06
- EIH_J           6
-   1.8267275721562653E-03   3.7694095940709932E-04  -2.5800198920720098E-14   4.3143175972322548E-13   1.8039111539361994E-13   2.1713576954972813E-01  -6.7004903028908230E-14   1.8104291246608195E-13   7.7082870081550862E-14  -4.0137572825887210E-14   8.9868548617298611E-14   3.8844352710840303E-14  -1.1185686650125868E-06   2.5044898177957639E-06   1.0825287304622227E-06
- EIH_J           7
-  -4.6069971388918335E-04   3.3002838374258107E-04  -9.5762975098236055E-15  -8.4973342605628653E-15  -3.6268418536049216E-15  -5.2324247056364581E-01  -6.0036422729217825E-15  -5.6348696678953341E-15  -2.0445804766718658E-15  -1.5592168472863633E-15  -1.0102356860488801E-15  -4.2040533898326204E-16  -6.1722817466729904E-07  -3.9990969156658773E-07  -1.6642073900922206E-07
- EIH_J           8
-  -1.3833207293088105E-05   3.6607732021977330E-04  -7.8158335457579833E-15   1.4187201103418543E-15   7.6165736549031286E-16  -1.1749593301702183E-01  -2.0914038379557998E-16  -4.5403758831260554E-16  -1.7670619613063517E-16  -6.5195481932820407E-16   1.2526726133306638E-16   6.7504052312475595E-17  -3.2314055476263021E-07   6.2088554483688428E-08   3.3458295370001441E-08
- EIH_J           9
-  -4.0183313337176072E-04   3.0895044349114230E-04  -2.0102868249379811E-19   5.7798792571367933E-19   2.5739492661180171E-19   9.1383803661908836E-01  -1.0269177216365394E-20   4.1340067653072541E-19   2.2277874851025063E-19  -2.3525758205104501E-20   5.1892338309511123E-20   2.3282022067309793E-20  -9.8318765845070454E-08   2.1686827752480254E-07   9.7300144636312985E-08
- EIH_J          10
-  -1.3600571884643616E-02   1.3291989968061181E-04   7.6872204179482010E-17  -3.9749697360802021E-17  -1.9551721384715176E-17   1.3908325584879361E-01   4.5672292598200297E-16   2.2444611507999467E-16   9.1066643447599176E-17  -6.3047884554668352E-16  -2.4053710442036868E-15  -9.3478747714848892E-16  -6.6685804896802220E-05  -2.5441631434770708E-04  -9.8872556567772065E-05
