--- examples/tests/vary_acc.out.direct
@@ -1,12 +1,0 @@
-  1    2.4588495000000000e+06    1.1878957764817864e-05   -5.1831723117613977e-06   -2.8420875102950962e-06
-  2    2.4588495000000000e+06   -5.1831741752338817e-06   -5.4360783927130500e-06    7.8810206990509273e-07
-  3    2.4588495000000000e+06   -2.8420869173720331e-06    7.8810198520179800e-07   -6.4428803461927036e-06
-  4    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  5    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  6    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    1.1878958719997790e-05   -5.1831726040988588e-06   -2.8420875459455623e-06
-  0    2.4588495000000000e+06   -5.1831726040988588e-06   -5.4360784024395903e-06    7.8810197748219764e-07
-  0    2.4588495000000000e+06   -2.8420875459455623e-06    7.8810197748219764e-07   -6.4428803175581958e-06
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
