--- examples/tests/vary_acc.out.earth_J2J4
@@ -1,12 +1,0 @@
-  1    2.4588495000000000e+06    4.1465342085381443e-24   -3.0860426577554260e-24   -6.1183130611164558e-24
-  2    2.4588495000000000e+06   -3.0860425108186322e-24   -2.2931973585963204e-25    3.1622467973781671e-24
-  3    2.4588495000000000e+06   -6.1183129876480588e-24    3.1622467973781671e-24   -3.9172143624759169e-24
-  4    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  5    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  6    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    4.1465340719147192e-24   -3.0860426192390132e-24   -6.1183130031138051e-24
-  0    2.4588495000000000e+06   -3.0860426192390132e-24   -2.2931964888142378e-25    3.1622468035906874e-24
-  0    2.4588495000000000e+06   -6.1183130031138051e-24    3.1622468035906874e-24   -3.9172144230332966e-24
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
