--- examples/tests/vary_acc.out.eih
@@ -1,12 +1,0 @@
-  1    2.4588495000000000e+06   -1.6974679676028930e-13    7.9451031050564674e-14    4.1365226341200990e-14
-  2    2.4588495000000000e+06    8.5120085913715573e-14    9.6207230587073979e-14    1.1293020014681175e-14
-  3    2.4588495000000000e+06    4.3683534871541206e-14    1.1510537049698834e-14    6.3604721037077198e-14
-  4    2.4588495000000000e+06    5.1407570903423452e-12    2.4357250723220546e-11    9.7967745959052818e-12
-  5    2.4588495000000000e+06   -1.3970269453121841e-11   -2.3554940513280292e-12   -7.6913980068676628e-13
-  6    2.4588495000000000e+06   -5.8797957697949717e-12   -2.2445200753959605e-12    1.4811880337231327e-13
-  0    2.4588495000000000e+06   -1.6974680355121597e-13    7.9451033642012556e-14    4.1365228252909014e-14
-  0    2.4588495000000000e+06    8.5120081784574373e-14    9.6207230534676714e-14    1.1293019784182437e-14
-  0    2.4588495000000000e+06    4.3683531593036218e-14    1.1510537526242287e-14    6.3604720374773172e-14
-  0    2.4588495000000000e+06    5.1407340956185834e-12    2.4357248608365544e-11    9.7967734361601048e-12
-  0    2.4588495000000000e+06   -1.3970261791558936e-11   -2.3554877029070265e-12   -7.6914096037994037e-13
-  0    2.4588495000000000e+06   -5.8797881096075740e-12   -2.2445221916542467e-12    1.4812228518233220e-13
