--- examples/tests/vary_acc.out.non_gravs
@@ -1,12 +1,0 @@
-  1    2.4588495000000000e+06   -1.0958603267918561e-10    6.3173498487976064e-11    2.1508794897183073e-11
-  2    2.4588495000000000e+06    5.5841978291538077e-11    4.6454626052161131e-11   -7.1384003421532772e-12
-  3    2.4588495000000000e+06    2.4111539374096488e-11   -8.7986038534529151e-12    6.3309150615438505e-11
-  4    2.4588495000000000e+06   -2.6396845536124437e-11   -3.7206775238151838e-11   -1.0622606832873040e-10
-  5    2.4588495000000000e+06    2.8036416771878044e-10    3.9517797023707382e-10    1.1282399441626127e-09
-  6    2.4588495000000000e+06   -6.8547093806728398e-10   -9.6618127699500874e-10   -2.7584710073447966e-09
-  0    2.4588495000000000e+06   -1.0958603162797791e-10    6.3173500054277179e-11    2.1508796229293504e-11
-  0    2.4588495000000000e+06    5.5841977684459388e-11    4.6454626614724946e-11   -7.1383985221426382e-12
-  0    2.4588495000000000e+06    2.4111548517408601e-11   -8.7986054758407031e-12    6.3309150575450139e-11
-  0    2.4588495000000000e+06   -2.6396854098655784e-11   -3.7206790042009810e-11   -1.0622610466648532e-10
-  0    2.4588495000000000e+06    2.8036440459927931e-10    3.9517813365910480e-10    1.1282412092142599e-09
-  0    2.4588495000000000e+06   -6.8547153323397366e-10   -9.6618314142630298e-10   -2.7584715422174517e-09
