--- examples/tests/vary_acc.out.simple_GR
@@ -1,12 +1,0 @@
-  1    2.4588495000000000e+06   -1.6962822276208117e-13    7.9439603927036004e-14    4.1353064314853016e-14
-  2    2.4588495000000000e+06    8.5137185893785867e-14    9.6206212009873157e-14    1.1331045950212087e-14
-  3    2.4588495000000000e+06    4.3683428848635545e-14    1.1550455935857914e-14    6.3528796961482021e-14
-  4    2.4588495000000000e+06    5.1504228502905470e-12    2.4340880503227540e-11    9.7895813431285637e-12
-  5    2.4588495000000000e+06   -1.3959493888311206e-11   -2.3631343389579615e-12   -7.7166929169905523e-13
-  6    2.4588495000000000e+06   -5.8756424436347095e-12   -2.2465787966236632e-12    1.4525410926863344e-13
-  0    2.4588495000000000e+06   -1.6962822375096908e-13    7.9439605545468010e-14    4.1353065603912642e-14
-  0    2.4588495000000000e+06    8.5137180621018196e-14    9.6206212745604331e-14    1.1331047064553685e-14
-  0    2.4588495000000000e+06    4.3683428232073126e-14    1.1550455239355455e-14    6.3528796429412014e-14
-  0    2.4588495000000000e+06    5.1503998620855509e-12    2.4340878384010046e-11    9.7895801817733507e-12
-  0    2.4588495000000000e+06   -1.3959486222835958e-11   -2.3631279804764735e-12   -7.7167045267678548e-13
-  0    2.4588495000000000e+06   -5.8756347809440524e-12   -2.2465809156721109e-12    1.4525759539732005e-13
