--- examples/tests/vary_acc.out.solar_J2
@@ -1,12 +1,0 @@
-  1    2.4588495000000000e+06    1.3910367678858232e-17   -5.9746261905283007e-18   -1.4730332660829889e-19
-  2    2.4588495000000000e+06   -5.9746272690490695e-18   -3.0536822589288878e-18    3.0495697436814399e-18
-  3    2.4588495000000000e+06   -1.4730282586651335e-19    3.0495697822000388e-18   -1.0856685997708328e-17
-  4    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  5    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  6    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    1.3910368541775220e-17   -5.9746267867018301e-18   -1.4730325023546837e-19
-  0    2.4588495000000000e+06   -5.9746267867018285e-18   -3.0536825636361907e-18    3.0495696947303797e-18
-  0    2.4588495000000000e+06   -1.4730325023546894e-19    3.0495696947303813e-18   -1.0856685978139034e-17
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
-  0    2.4588495000000000e+06    0.0000000000000000e+00    0.0000000000000000e+00    0.0000000000000000e+00
